+++ CHANGELOG.md
@@ -156,7 +156,7 @@ The minimum Omeka S version required is 3.0.0
 
 First release
 
-[0.12.1]: https://github.com/biblibre/omeka-s-module-Search/compare/v0.12.1...v0.13.0
+[0.13.0]: https://github.com/biblibre/omeka-s-module-Search/compare/v0.12.1...v0.13.0
 [0.12.1]: https://github.com/biblibre/omeka-s-module-Search/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/biblibre/omeka-s-module-Search/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/biblibre/omeka-s-module-Search/compare/v0.10.0...v0.11.0